Move 'datastore.api' functions to d'atastore.client.Client 'methods.
Remove the 'datastore.api' module altogether.

See #944.
