+++ gcloud/datastore/__init__.py
@@ -54,13 +54,6 @@ from gcloud.datastore._implicit_environ import get_default_connection
 from gcloud.datastore._implicit_environ import get_default_dataset_id
 from gcloud.datastore._implicit_environ import set_default_connection
 from gcloud.datastore._implicit_environ import set_default_dataset_id
-from gcloud.datastore.api import allocate_ids
-from gcloud.datastore.api import delete
-from gcloud.datastore.api import delete_multi
-from gcloud.datastore.api import get
-from gcloud.datastore.api import get_multi
-from gcloud.datastore.api import put
-from gcloud.datastore.api import put_multi
 from gcloud.datastore.batch import Batch
 from gcloud.datastore.connection import SCOPE
 from gcloud.datastore.connection import Connection
