+++ system_tests/datastore.py
@@ -32,16 +32,18 @@ class TestDatastore(unittest2.TestCase):
         self.case_entities_to_delete = []
 
     def tearDown(self):
-        with datastore.Transaction():
+        client = datastore.Client()
+        with client.transaction():
             keys = [entity.key for entity in self.case_entities_to_delete]
-            datastore.delete_multi(keys)
+            client.delete_multi(keys)
 
 
 class TestDatastoreAllocateIDs(TestDatastore):
 
     def test_allocate_ids(self):
         num_ids = 10
-        allocated_keys = datastore.allocate_ids(datastore.Key('Kind'), num_ids)
+        client = datastore.Client()
+        allocated_keys = client.allocate_ids(client.key('Kind'), num_ids)
         self.assertEqual(len(allocated_keys), num_ids)
 
         unique_ids = set()
@@ -82,7 +84,8 @@ class TestDatastoreSave(TestDatastore):
 
     def _generic_test_post(self, name=None, key_id=None):
         entity = self._get_post(id_or_name=(name or key_id))
-        datastore.put(entity)
+        client = datastore.Client()
+        client.put(entity)
 
         # Register entity to be deleted.
         self.case_entities_to_delete.append(entity)
@@ -91,7 +94,7 @@ class TestDatastoreSave(TestDatastore):
             self.assertEqual(entity.key.name, name)
         if key_id is not None:
             self.assertEqual(entity.key.id, key_id)
-        retrieved_entity = datastore.get(entity.key)
+        retrieved_entity = client.get(entity.key)
         # Check the given and retrieved are the the same.
         self.assertEqual(retrieved_entity, entity)
 
@@ -126,7 +129,8 @@ class TestDatastoreSave(TestDatastore):
             self.case_entities_to_delete.append(entity2)
 
         keys = [entity1.key, entity2.key]
-        matches = datastore.get_multi(keys)
+        client = datastore.Client()
+        matches = client.get_multi(keys)
         self.assertEqual(len(matches), 2)
 
     def test_empty_kind(self):
@@ -145,7 +149,8 @@ class TestDatastoreSaveKeys(TestDatastore):
         entity['fullName'] = u'Full name'
         entity['linkedTo'] = key  # Self reference.
 
-        datastore.put(entity)
+        client = datastore.Client()
+        client.put(entity)
         self.case_entities_to_delete.append(entity)
 
         query = datastore.Query(kind='Person')
@@ -326,16 +331,17 @@ class TestDatastoreQuery(TestDatastore):
 class TestDatastoreTransaction(TestDatastore):
 
     def test_transaction(self):
+        client = datastore.Client()
         entity = datastore.Entity(key=datastore.Key('Company', 'Google'))
         entity['url'] = u'www.google.com'
 
         with datastore.Transaction() as xact:
-            result = datastore.get(entity.key)
+            result = client.get(entity.key)
             if result is None:
                 xact.put(entity)
                 self.case_entities_to_delete.append(entity)
 
         # This will always return after the transaction.
-        retrieved_entity = datastore.get(entity.key)
+        retrieved_entity = client.get(entity.key)
         self.case_entities_to_delete.append(retrieved_entity)
         self.assertEqual(retrieved_entity, entity)
